Trait ckb_rpc::module::ExperimentRpc

pub trait ExperimentRpc: Sized + Send + Sync + 'static {
    fn compute_transaction_hash(&self, tx: Transaction) -> Result<H256>;

    fn compute_script_hash(&self, script: Script) -> Result<H256>;

    fn dry_run_transaction(&self, tx: Transaction) -> Result<DryRunResult>;

    fn calculate_dao_maximum_withdraw(
        &self, 
        out_point: OutPoint, 
        block_hash: H256
    ) -> Result<Capacity>;

    fn estimate_fee_rate(
        &self, 
        expect_confirm_blocks: Uint64
    ) -> Result<EstimateResult>;

    fn to_delegate<M: Metadata>(self) -> IoDelegate<Self, M> { ... }
}

RPC Module Experiment for experimenting methods.

EXPERIMENTAL warning

The methods here may be removed or changed in future releases without prior notifications.

Required methods

fn compute_transaction_hash(&self, tx: Transaction) -> Result<H256>

👎 Deprecated since 0.36.0:
Please implement molecule and compute the transaction hash in clients.

Returns the transaction hash for the given transaction.

Examples

Request

{
  "id": 42,
  "jsonrpc": "2.0",
  "method": "_compute_transaction_hash",
  "params": [
    {
      "cell_deps": [
        {
          "dep_type": "code",
          "out_point": {
            "index": "0x0",
            "tx_hash": "0xa4037a893eb48e18ed4ef61034ce26eba9c585f15c9cee102ae58505565eccc3"
          }
        }
      ],
      "header_deps": [
        "0x7978ec7ce5b507cfb52e149e36b1a23f6062ed150503c85bbf825da3599095ed"
      ],
      "inputs": [
        {
          "previous_output": {
            "index": "0x0",
            "tx_hash": "0x365698b50ca0da75dca2c87f9e7b563811d3b5813736b8cc62cc3b106faceb17"
          },
          "since": "0x0"
        }
      ],
      "outputs": [
        {
          "capacity": "0x2540be400",
          "lock": {
            "args": "0x",
            "code_hash": "0x28e83a1277d48add8e72fadaa9248559e1b632bab2bd60b27955ebc4c03800a5",
            "hash_type": "data"
          },
          "type": null
        }
      ],
      "outputs_data": [
        "0x"
      ],
      "version": "0x0",
      "witnesses": []
    }
  ]
}

Response

{
  "id": 42,
  "jsonrpc": "2.0",
  "result": "0xa0ef4eb5f4ceeb08a4c8524d84c5da95dce2f608e0ca2ec8091191b0f330c6e3"
}

fn compute_script_hash(&self, script: Script) -> Result<H256>

👎 Deprecated since 0.36.0:
Please implement molecule and compute the script hash in clients.

Returns the script hash for the given script.

Examples

Request

{
  "id": 42,
  "jsonrpc": "2.0",
  "method": "_compute_script_hash",
  "params": [
    {
      "args": "0x",
      "code_hash": "0x28e83a1277d48add8e72fadaa9248559e1b632bab2bd60b27955ebc4c03800a5",
      "hash_type": "data"
    }
  ]
}

Response

{
  "id": 42,
  "jsonrpc": "2.0",
  "result": "0x4ceaa32f692948413e213ce6f3a83337145bde6e11fd8cb94377ce2637dcc412"
}

fn dry_run_transaction(&self, tx: Transaction) -> Result<DryRunResult>

Dry run transaction and return the execution cycles.

This method will not check the transaction validity, but only run the lock script and type script and then return the execution cycles.

It is used to debug transaction scripts and query how many cycles the scripts consume.

Examples

Request

{
  "id": 42,
  "jsonrpc": "2.0",
  "method": "dry_run_transaction",
  "params": [
    {
      "cell_deps": [
        {
          "dep_type": "code",
          "out_point": {
            "index": "0x0",
            "tx_hash": "0xa4037a893eb48e18ed4ef61034ce26eba9c585f15c9cee102ae58505565eccc3"
          }
        }
      ],
      "header_deps": [
        "0x7978ec7ce5b507cfb52e149e36b1a23f6062ed150503c85bbf825da3599095ed"
      ],
      "inputs": [
        {
          "previous_output": {
            "index": "0x0",
            "tx_hash": "0x365698b50ca0da75dca2c87f9e7b563811d3b5813736b8cc62cc3b106faceb17"
          },
          "since": "0x0"
        }
      ],
      "outputs": [
        {
          "capacity": "0x2540be400",
          "lock": {
            "args": "0x",
            "code_hash": "0x28e83a1277d48add8e72fadaa9248559e1b632bab2bd60b27955ebc4c03800a5",
            "hash_type": "data"
          },
          "type": null
        }
      ],
      "outputs_data": [
        "0x"
      ],
      "version": "0x0",
      "witnesses": []
    }
  ]
}

Response

{
  "id": 42,
  "jsonrpc": "2.0",
  "result": {
    "cycles": "0x219"
  }
}

fn calculate_dao_maximum_withdraw(
&self,
out_point: OutPoint,
block_hash: H256
) -> Result<Capacity>

Calculates the maximum withdraw one can get, given a referenced DAO cell, and a withdraw block hash.

Params

out_point - Reference to the DAO cell.

block_hash - The assumed reference block for withdraw. This block must be in the canonical chain.

Returns

The RPC returns the final capacity when the cell out_point is withdrawn using the block block_hash as the reference.

In CKB, scripts cannot get the information about in which block the transaction is committed. A workaround is letting the transaction reference a block hash so the script knows that the transaction is committed at least after the reference block.

Examples

Request

{
  "id": 42,
  "jsonrpc": "2.0",
  "method": "calculate_dao_maximum_withdraw",
  "params": [
    {
      "index": "0x0",
      "tx_hash": "0xa4037a893eb48e18ed4ef61034ce26eba9c585f15c9cee102ae58505565eccc3"
    },
    "0xa5f5c85987a15de25661e5a214f2c1449cd803f071acc7999820f25246471f40"
  ]
}

Response

{
  "id": 42,
  "jsonrpc": "2.0",
  "result": "0x4a8b4e8a4"
}

fn estimate_fee_rate(
&self,
expect_confirm_blocks: Uint64
) -> Result<EstimateResult>

👎 Deprecated since 0.34.0:
This method is deprecated because of the performance issue. It always returns an error now.

Estimates a fee rate (capacity/KB) for a transaction that to be committed within the expect number of blocks.

Loading content...
Provided methods

fn to_delegate<M: Metadata>(self) -> IoDelegate<Self, M>

Create an IoDelegate, wiring rpc calls to the trait methods.

Loading content...