Trait ckb_rpc::module::MinerRpc

pub trait MinerRpc: Sized + Send + Sync + 'static {
    fn get_block_template(
        &self, 
        bytes_limit: Option<Uint64>, 
        proposals_limit: Option<Uint64>, 
        max_version: Option<Version>
    ) -> Result<BlockTemplate>;

    fn submit_block(&self, _work_id: String, _data: Block) -> Result<H256>;

    fn to_delegate<M: Metadata>(self) -> IoDelegate<Self, M> { ... }
}

Required methods

fn get_block_template(
&self,
bytes_limit: Option<Uint64>,
proposals_limit: Option<Uint64>,
max_version: Option<Version>
) -> Result<BlockTemplate>

Returns data needed to construct a block to work on

Params

bytes_limit - the max serialization size in bytes of the block. (Optional: the default is the consensus limit.)

proposals_limit - the max count of proposals. (Optional: the default is the consensus limit.)

max_version - the max block version. (Optional: the default is one configured in the current client version.) max version.

Examples

Request

{
  "jsonrpc": "2.0",
  "method": "get_block_template",
  "params": [
    null,
    null,
    null
  ],
  "id": 42
}

Response

{
  "jsonrpc": "2.0",
  "result": {
    "bytes_limit": "0x22d387",
    "cellbase": {
        "cycles": null,
    }
  },
  "id": 42
}

fn submit_block(&self, _work_id: String, _data: Block) -> Result<H256>

Loading content...
Provided methods

fn to_delegate<M: Metadata>(self) -> IoDelegate<Self, M>

Create an IoDelegate, wiring rpc calls to the trait methods.

Loading content...