Trait ckb_rpc::module::SubscriptionRpc

pub trait SubscriptionRpc: Sized + Send + Sync + 'static {
    type Metadata: PubSubMetadata;
    fn subscribe(
        &self, 
        meta: Self::Metadata, 
        subscriber: Subscriber<String>, 
        topic: Topic
    );

    fn unsubscribe(
        &self, 
        meta: Option<Self::Metadata>, 
        id: SubscriptionId
    ) -> Result<bool>;

    fn to_delegate(self) -> IoDelegate<Self, Self::Metadata> { ... }
}

RPC Module Subscription that CKB node will push new messages to subscribers.

RPC subscriptions require a full duplex connection. CKB offers such connections in the form of TCP (enable with rpc.tcp_listen_address configuration option) and WebSocket (enable with rpc.ws_listen_address).

Examples

TCP RPC subscription:

telnet localhost 18114
> {"id": 2, "jsonrpc": "2.0", "method": "subscribe", "params": ["new_tip_header"]}
< {"jsonrpc":"2.0","result":0,"id":2}
< {"jsonrpc":"2.0","method":"subscribe","params":{"result":"...block header json...",
"subscription":0}}
< {"jsonrpc":"2.0","method":"subscribe","params":{"result":"...block header json...",
"subscription":0}}
< ...
> {"id": 2, "jsonrpc": "2.0", "method": "unsubscribe", "params": [0]}
< {"jsonrpc":"2.0","result":true,"id":2}

WebSocket RPC subscription:

let socket = new WebSocket("ws://localhost:28114")

socket.onmessage = function(event) {
  console.log(`Data received from server: ${event.data}`);
}

socket.send(`{"id": 2, "jsonrpc": "2.0", "method": "subscribe", "params": ["new_tip_header"]}`)

socket.send(`{"id": 2, "jsonrpc": "2.0", "method": "unsubscribe", "params": [0]}`)

Associated Types

type Metadata: PubSubMetadata

Loading content...
Required methods

fn subscribe(
&self,
meta: Self::Metadata,
subscriber: Subscriber<String>,
topic: Topic
)

Subscribes to a topic.

Params

topic - Subscription topic (enum: new_tip_header | new_tip_block)

Returns

This RPC returns the subscription ID as the result. CKB node will push messages in the subscribed topics to current RPC connection. The subscript ID is also attached as params.subscription in the push messages.

Topics

new_tip_header

Whenever there's a block is appended to the canonical chain, CKB node will publish the block header to subscribers.

The schema of the params.result in the notification is the same with the response result of RPC get_header when verbosity is 2.

Example:

{
  "jsonrpc": "2.0",
  "method": "subscribe",
  "params": {
    "result": {
      "compact_target": "0x1e083126",
      "dao": "0xb5a3e047474401001bc476b9ee573000c0c387962a38000000febffacf030000",
      "epoch": "0x7080018000001",
      "hash": "0xa5f5c85987a15de25661e5a214f2c1449cd803f071acc7999820f25246471f40",
      "nonce": "0x0",
      "number": "0x400",
      "parent_hash": "0xae003585fa15309b30b31aed3dcf385e9472c3c3e93746a6c4540629a6a1ed2d",
      "proposals_hash": "0x0000000000000000000000000000000000000000000000000000000000000000",
      "timestamp": "0x5cd2b117",
      "transactions_root": "0xc47d5b78b3c4c4c853e2a32810818940d0ee403423bea9ec7b8e566d9595206c",
      "uncles_hash": "0x0000000000000000000000000000000000000000000000000000000000000000",
      "version": "0x0"
    },
    "subscription": "0x2a"
  }
}

new_tip_block

Whenever there's a block is appended to the canonical chain, CKB node will publish the whole block to subscribers.

The schema of the params.result in the notification is the same with the response result of RPC get_block when verbosity is 2.

Example:

{
  "jsonrpc": "2.0",
  "method": "subscribe",
  "params": {
    "result": {
      "header": {
        "compact_target": "0x1e083126",
        "dao": "0xb5a3e047474401001bc476b9ee573000c0c387962a38000000febffacf030000",
        "epoch": "0x7080018000001",
        "hash": "0xa5f5c85987a15de25661e5a214f2c1449cd803f071acc7999820f25246471f40",
        "nonce": "0x0",
        "number": "0x400",
        "parent_hash": "0xae003585fa15309b30b31aed3dcf385e9472c3c3e93746a6c4540629a6a1ed2d",
        "proposals_hash": "0x0000000000000000000000000000000000000000000000000000000000000000",
        "timestamp": "0x5cd2b117",
        "transactions_root": "0xc47d5b78b3c4c4c853e2a32810818940d0ee403423bea9ec7b8e566d9595206c",
        "uncles_hash": "0x0000000000000000000000000000000000000000000000000000000000000000",
        "version": "0x0"
      },
      "proposals": [],
      "transactions": [
        {
          "cell_deps": [],
          "hash": "0x365698b50ca0da75dca2c87f9e7b563811d3b5813736b8cc62cc3b106faceb17",
          "header_deps": [],
          "inputs": [
            {
              "previous_output": {
                "index": "0xffffffff",
                "tx_hash": "0x0000000000000000000000000000000000000000000000000000000000000000"
              },
              "since": "0x400"
            }
          ],
          "outputs": [
            {
              "capacity": "0x18e64b61cf",
              "lock": {
                "args": "0x",
                "code_hash": "0x28e83a1277d48add8e72fadaa9248559e1b632bab2bd60b27955ebc4c03800a5",
                "hash_type": "data"
              },
              "type": null
            }
          ],
          "outputs_data": [
            "0x"
          ],
          "version": "0x0",
          "witnesses": [
            "0x450000000c000000410000003500000010000000300000003100000028e83a1277d48add8e72fadaa9248559e1b632bab2bd60b27955ebc4c03800a5000000000000000000"
          ]
        }
      ],
      "uncles": []
    },
    "subscription": "0x2a"
  }
}

Examples

Request

{
  "id": 42,
  "jsonrpc": "2.0",
  "method": "subscribe",
  "params": [
    "new_tip_header"
  ]
}

Response

{
  "id": 42,
  "jsonrpc": "2.0",
  "result": "0x2a"
}

fn unsubscribe(
&self,
meta: Option<Self::Metadata>,
id: SubscriptionId
) -> Result<bool>

Unsubscribes from a subscribed topic.

Params

id - Subscription ID

Examples

Request

{
  "id": 42,
  "jsonrpc": "2.0",
  "method": "unsubscribe",
  "params": [
    "0x2a"
  ]
}

Response

{
  "id": 42,
  "jsonrpc": "2.0",
  "result": true
}

Loading content...
Provided methods

fn to_delegate(self) -> IoDelegate<Self, Self::Metadata>

Create an IoDelegate, wiring rpc calls to the trait methods.

Loading content...